+++ bucket.h
@@ -21,9 +21,9 @@ typedef struct Table {
 
 // used when iterating through a table's buckets
 bool valid_bucket(Table *table, int index);
-void *table_search(Table *table, int key);
-void table_insert(Table *table, int key, void *data);
-void table_delete_item(Table *table, int key);
+void *table_search(Table *table, unsigned long key);
+void table_insert(Table *table, unsigned long key, void *data);
+void table_delete_item(Table *table, unsigned long key);
 void table_free(Table *table);
 Table *table_init(int size);
 void *table_iterate(Table *table, bool first_iterate);
